Buffer: Optimise to minimise memmove shuffles.
Previously the data in the linear buffer was always stored at
the start of the allocation, pointed to by `buffer->data`.
This was achieved by memmoving every time data was consumed
from the front.

Now the allocation is pointed to by `buffer->alloc`, and the
start of the data is pointed to by `buffer->data` (as before).

This means client code does not need to change to get at the
data.

The advantage comes when we discard the from the start of the
buffer, when some data is consumed.  We now simply advance the
data pointer by the number of bytes to be discarded, and reduce
the buffer length by the same amount.

If the used portion of the buffer now fits between the start
of the allocation and the current start of the data, it is
memcpyed to the allocation start, otherwise it is left alone.

This is a significant optimisation when the size of the chunk
is large, such as when loading from disc. (When the first
(only) "chunk" is just the whole file.